Managin Proper Status Codes Update

+++ src/main/java/com/munusync/backend/service/UserService.java
@@ -3,7 +3,9 @@ package com.munusync.backend.service;
 import java.util.List;
 import java.util.Optional;
 
+import org.apache.catalina.connector.Response;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.munusync.backend.entity.User;
@@ -24,11 +26,11 @@ public class UserService {
         return userRepository.findById(id);
     }
 
-    public void createUser(User user){
-        this.userRepository.save(user);
+    public User createUser(User user){
+        return userRepository.save(user) ;
     }
 
-    public void updateUser(Long id, User user){
+    public ResponseEntity<?> updateUser(Long id, User user){
 
         Optional<User> optionalUser = getUserById(id);
 
@@ -38,13 +40,21 @@ public class UserService {
 
             userRefrence.setName(user.getName());
             userRefrence.setEmail(user.getEmail());
+
+            User updateUser = userRepository.save(userRefrence);
            
-            userRepository.save(userRefrence);
+           return ResponseEntity.ok(updateUser);
         }
+        return ResponseEntity.notFound().build();
     }
 
-    public void deleteUser(Long id){
-        userRepository.deleteById(id);
+    public ResponseEntity<?> deleteUser(Long id){
+        if (this.getUserById(id).isPresent()) {
+            userRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }else{
+            return ResponseEntity.notFound().build();
+        }
     }
     
 }
